Added a lot of classes

+++ src/main/java/software/anton/pcep/functions/IncomingWindowFunction.java
@@ -12,17 +12,17 @@ import static software.anton.pcep.configs.Configuration.GRAFANA_PANEL;
 /**
  * @author Anton Rudacov <anton.rudacov @ gmail.com>
  */
-public class IncomingWindowFunction implements WindowFunction<KeyedDataPoint<Integer>, Integer, String, GlobalWindow> {
+public class IncomingWindowFunction implements WindowFunction<KeyedDataPoint<Double>, Double, String, GlobalWindow> {
 
     private final static GrafanaAnnotator ANNOTATOR_IN = new GrafanaAnnotator(GRAFANA_DASHBOARD, GRAFANA_PANEL);
     private final static GrafanaAnnotator ANNOTATOR_BOTH = new GrafanaAnnotator(GRAFANA_DASHBOARD, 4);
 
     @Override
-    public void apply(String key, GlobalWindow window, Iterable<KeyedDataPoint<Integer>> input, Collector<Integer> out) throws Exception {
+    public void apply(String key, GlobalWindow window, Iterable<KeyedDataPoint<Double>> input, Collector<Double> out) throws Exception {
 
-        int sum = 0;
+        double sum = 0;
         long lastTimeStamp = 0L;
-        for (KeyedDataPoint<Integer> point : input) {
+        for (KeyedDataPoint<Double> point : input) {
             System.out.println(point);
             sum += point.getValue();
             lastTimeStamp = point.getTimeStamp();
