+++ src/main/java/software/anton/pcep/functions/ModelTrainerFunction.java
@@ -9,10 +9,10 @@ import software.anton.pcep.data.KeyedDataPoint;
 /**
  * @author Anton Rudacov <anton.rudacov @ gmail.com>
  */
-public class ModelTrainerFunction implements WindowFunction<KeyedDataPoint<Integer>, Integer, String, GlobalWindow> {
+public class ModelTrainerFunction implements WindowFunction<KeyedDataPoint<Double>, Double, String, GlobalWindow> {
 
     @Override
-    public void apply(String s, GlobalWindow window, Iterable<KeyedDataPoint<Integer>> input, Collector<Integer> out) throws Exception {
+    public void apply(String s, GlobalWindow window, Iterable<KeyedDataPoint<Double>> input, Collector<Double> out) throws Exception {
 
         // out.collect();
     }
